Cleaned up Mapper and Composite Mapper.
Mapper and Composite Mapper are now much simpler, and do not implement the load and save functions. This is considered too implementation specific to be kept in the base class. Subclasses will provide these - and other - functions.

I also deleted MapperInterface and CompositeMapperInterface in favour of abstract base classes. I'm not sure if you can have a static abstract base class, so Mapper implements its load/save functions and just throws an exception if they're called. That said, it shouldn't be possible to call them, since you can't instantiate the base class.